requirements.txt
functions-framework==3.*
gspread==6.2.0
google-cloud-storage==3.1.0
google-cloud-secret-manager==2.23.2
google-cloud-bigquery==3.30.0
